+++ src/Title/Title.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Text, View } from 'react-native';
+import { Text, View, TouchableOpacity } from 'react-native';
 import PropTypes from 'prop-types';
 
 import { getCurrentMonth, availableNumberOfDays } from '../utils';
@@ -12,23 +12,31 @@ const getFontSizeHeader = (numberOfDays) => {
   return 16;
 };
 
-const Title = ({ style, showTitle, numberOfDays, selectedDate, textStyle }) => {
+const Title = ({
+  style, showTitle, numberOfDays, selectedDate, textStyle, onMonthPress, width,
+}) => {
+  if (!showTitle) {
+    return <View style={[styles.title, style]}></View>
+  }
+  const formattedMonth = getCurrentMonth(selectedDate);
   return (
-    <View style={[styles.title, style]}>
-      {showTitle ? (
-        <Text
-          style={[
-            {
-              fontSize: getFontSizeHeader(numberOfDays),
-              textAlign: 'center',
-            },
-            textStyle,
-          ]}
-        >
-          {getCurrentMonth(selectedDate)}
-        </Text>
-      ) : null}
-    </View>
+    <TouchableOpacity
+      style={[styles.title, { width }, style]}
+      onPress={() => onMonthPress && onMonthPress(selectedDate, formattedMonth)}
+      disabled={!onMonthPress}
+    >
+      <Text
+        style={[
+          {
+            fontSize: getFontSizeHeader(numberOfDays),
+            textAlign: 'center',
+          },
+          textStyle,
+        ]}
+      >
+        {formattedMonth}
+      </Text>
+    </TouchableOpacity>
   );
 };
 
@@ -38,6 +46,8 @@ Title.propTypes = {
   selectedDate: PropTypes.instanceOf(Date).isRequired,
   style: PropTypes.object,
   textStyle: PropTypes.object,
+  onMonthPress: PropTypes.func,
+  width: PropTypes.number.isRequired,
 };
 
 export default React.memo(Title);