+++ src/Title/Title.styles.js
@@ -4,7 +4,6 @@ const styles = StyleSheet.create({
   title: {
     justifyContent: 'center',
     alignItems: 'center',
-    width: 60,
     borderTopWidth: 1,
   },
 });