+++ src/WeekView/WeekView.js
@@ -19,14 +19,30 @@ import Title from '../Title/Title';
 import Times from '../Times/Times';
 import styles from './WeekView.styles';
 import {
-  CONTAINER_HEIGHT,
   DATE_STR_FORMAT,
   availableNumberOfDays,
   setLocale,
-  CONTAINER_WIDTH,
+  minutesToYDimension,
+  computeWeekViewDimensions,
 } from '../utils';
 
 const MINUTES_IN_DAY = 60 * 24;
+const calculateTimesArray = (
+  minutesStep, formatTimeLabel, beginAt = 0, endAt = MINUTES_IN_DAY,
+) => {
+  const times = [];
+  const startOfDay = moment().startOf('day');
+  for (
+    let timer = (0 <= beginAt && beginAt < MINUTES_IN_DAY) ? beginAt : 0;
+    timer < endAt && timer < MINUTES_IN_DAY;
+    timer += minutesStep
+  ) {
+    const time = startOfDay.clone().minutes(timer);
+    times.push(time.format(formatTimeLabel));
+  }
+
+  return times;
+};
 
 export default class WeekView extends Component {
   constructor(props) {
@@ -94,21 +110,14 @@ export default class WeekView extends Component {
     this.eventsGridScrollX.removeAllListeners();
   }
 
-  calculateTimes = memoizeOne((minutesStep, formatTimeLabel) => {
-    const times = [];
-    const startOfDay = moment().startOf('day');
-    for (let timer = 0; timer < MINUTES_IN_DAY; timer += minutesStep) {
-      const time = startOfDay.clone().minutes(timer);
-      times.push(time.format(formatTimeLabel));
-    }
-    return times;
-  });
+  calculateTimes = memoizeOne(calculateTimesArray);
 
   scrollToVerticalStart = () => {
     if (this.verticalAgenda) {
-      const { startHour, hoursInDisplay } = this.props;
-      const startHeight = (startHour * CONTAINER_HEIGHT) / hoursInDisplay;
-      this.verticalAgenda.scrollTo({ y: startHeight, x: 0, animated: false });
+      const { startHour, hoursInDisplay, beginAgendaAt } = this.props;
+      const startHeight = minutesToYDimension(hoursInDisplay, startHour * 60);
+      const agendaOffset = minutesToYDimension(hoursInDisplay, beginAgendaAt);
+      this.verticalAgenda.scrollTo({ y: startHeight - agendaOffset, x: 0, animated: false });
     }
   };
 
@@ -358,11 +367,16 @@ export default class WeekView extends Component {
     return sortedEvents;
   });
 
-  getListItemLayout = (index) => ({
-    length: CONTAINER_WIDTH,
-    offset: CONTAINER_WIDTH * index,
-    index,
-  });
+  updateDimensions = memoizeOne(computeWeekViewDimensions);
+
+  getListItemLayout = (item, index) => {
+    const pageWidth = this.dimensions.pageWidth || 0;
+    return {
+      length: pageWidth,
+      offset: pageWidth * index,
+      index,
+    };
+  };
 
   render() {
     const {
@@ -374,6 +388,7 @@ export default class WeekView extends Component {
       gridRowStyle,
       gridColumnStyle,
       eventContainerStyle,
+      DayHeaderComponent,
       TodayHeaderComponent,
       formatDateHeader,
       onEventPress,
@@ -381,6 +396,8 @@ export default class WeekView extends Component {
       events,
       hoursInDisplay,
       timeStep,
+      beginAgendaAt,
+      endAgendaAt,
       formatTimeLabel,
       onGridClick,
       onGridLongPress,
@@ -391,15 +408,24 @@ export default class WeekView extends Component {
       showNowLine,
       nowLineColor,
       onDragEvent,
+      onMonthPress,
+      onDayPress,
       isRefreshing,
       RefreshComponent,
     } = this.props;
     const { currentMoment, initialDates } = this.state;
-    const times = this.calculateTimes(timeStep, formatTimeLabel);
+    const times = this.calculateTimes(timeStep, formatTimeLabel, beginAgendaAt, endAgendaAt);
     const eventsByDate = this.sortEventsByDate(events);
     const horizontalInverted =
       (prependMostRecent && !rightToLeft) ||
       (!prependMostRecent && rightToLeft);
+
+    this.dimensions = this.updateDimensions(numberOfDays);
+    const {
+      pageWidth,
+      dayWidth,
+      timeLabelsWidth,
+    } = this.dimensions;
 
     return (
       <GestureHandlerRootView style={styles.container}>
@@ -410,6 +436,8 @@ export default class WeekView extends Component {
             textStyle={headerTextStyle}
             numberOfDays={numberOfDays}
             selectedDate={currentMoment}
+            onMonthPress={onMonthPress}
+            width={timeLabelsWidth}
           />
           <VirtualizedList
             horizontal
@@ -421,20 +449,22 @@ export default class WeekView extends Component {
             data={initialDates}
             getItem={(data, index) => data[index]}
             getItemCount={(data) => data.length}
-            getItemLayout={(_, index) => this.getListItemLayout(index)}
+            getItemLayout={this.getListItemLayout}
             keyExtractor={(item) => item}
             initialScrollIndex={this.pageOffset}
             renderItem={({ item }) => {
               return (
-                <View key={item} style={styles.header}>
+                <View key={item} style={[styles.header, { width: pageWidth }]}>
                   <Header
                     style={headerStyle}
                     textStyle={headerTextStyle}
                     TodayComponent={TodayHeaderComponent}
+                    DayComponent={DayHeaderComponent}
                     formatDate={formatDateHeader}
                     initialDate={item}
                     numberOfDays={numberOfDays}
                     rightToLeft={rightToLeft}
+                    onDayPress={onDayPress}
                   />
                 </View>
               );
@@ -442,7 +472,9 @@ export default class WeekView extends Component {
           />
         </View>
         {isRefreshing && RefreshComponent && (
-          <RefreshComponent style={styles.loadingSpinner} />
+          <RefreshComponent
+            style={[styles.loadingSpinner, { right: pageWidth / 2 }]}
+          />
         )}
         <ScrollView
           onStartShouldSetResponderCapture={() => false}
@@ -455,12 +487,13 @@ export default class WeekView extends Component {
               textStyle={hourTextStyle}
               hoursInDisplay={hoursInDisplay}
               timeStep={timeStep}
+              width={timeLabelsWidth}
             />
             <VirtualizedList
               data={initialDates}
               getItem={(data, index) => data[index]}
               getItemCount={(data) => data.length}
-              getItemLayout={(_, index) => this.getListItemLayout(index)}
+              getItemLayout={this.getListItemLayout}
               keyExtractor={(item) => item}
               initialScrollIndex={this.pageOffset}
               scrollEnabled={!fixedHorizontally}
@@ -480,6 +513,7 @@ export default class WeekView extends Component {
                     onGridLongPress={onGridLongPress}
                     hoursInDisplay={hoursInDisplay}
                     timeStep={timeStep}
+                    beginAgendaAt={beginAgendaAt}
                     EventComponent={EventComponent}
                     eventContainerStyle={eventContainerStyle}
                     gridRowStyle={gridRowStyle}
@@ -488,6 +522,8 @@ export default class WeekView extends Component {
                     showNowLine={showNowLine}
                     nowLineColor={nowLineColor}
                     onDragEvent={onDragEvent}
+                    pageWidth={pageWidth}
+                    dayWidth={dayWidth}
                   />
                 );
               }}
@@ -539,9 +575,12 @@ WeekView.propTypes = {
   locale: PropTypes.string,
   hoursInDisplay: PropTypes.number,
   timeStep: PropTypes.number,
+  beginAgendaAt: PropTypes.number,
+  endAgendaAt: PropTypes.number,
   formatTimeLabel: PropTypes.string,
   startHour: PropTypes.number,
   EventComponent: PropTypes.elementType,
+  DayHeaderComponent: PropTypes.elementType,
   TodayHeaderComponent: PropTypes.elementType,
   showTitle: PropTypes.bool,
   rightToLeft: PropTypes.bool,
@@ -550,6 +589,8 @@ WeekView.propTypes = {
   showNowLine: PropTypes.bool,
   nowLineColor: PropTypes.string,
   onDragEvent: PropTypes.func,
+  onMonthPress: PropTypes.func,
+  onDayPress: PropTypes.func,
   isRefreshing: PropTypes.bool,
   RefreshComponent: PropTypes.elementType,
 };
@@ -560,6 +601,8 @@ WeekView.defaultProps = {
   hoursInDisplay: 6,
   weekStartsOn: 1,
   timeStep: 60,
+  beginAgendaAt: 0,
+  endAgendaAt: MINUTES_IN_DAY,
   formatTimeLabel: 'H:mm',
   startHour: 8,
   showTitle: true,