+++ src/WeekView/WeekView.styles.js
@@ -1,5 +1,5 @@
 import { StyleSheet } from 'react-native';
-import { CONTAINER_WIDTH, CONTAINER_HEIGHT } from '../utils';
+import { CONTAINER_HEIGHT } from '../utils';
 
 const styles = StyleSheet.create({
   container: {
@@ -16,12 +16,10 @@ const styles = StyleSheet.create({
     height: 50,
     justifyContent: 'center',
     alignItems: 'center',
-    width: CONTAINER_WIDTH,
   },
   loadingSpinner: {
     position: 'absolute',
     top: CONTAINER_HEIGHT / 2,
-    right: CONTAINER_WIDTH / 2,
     zIndex: 2,
   },
 });