+++ docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
-
+## 0.12.0 - 2022-04-03
+- Add beginAgendaAt and endAgendaAt props [#183](https://github.com/hoangnm/react-native-week-view/pull/183)
+## 0.11.0 - 2022-04-03
+- Add DayHeaderComponent prop [#182](https://github.com/hoangnm/react-native-week-view/pull/182)
+## 0.10.0 - 2022-03-27
+- Add onDayPress and onMonthPress props [#176](https://github.com/hoangnm/react-native-week-view/pull/176)
+- Add support for React 17 [#173](https://github.com/hoangnm/react-native-week-view/pull/173)
 ## 0.9.0 - 2022-01-18
 - Add gridRowStyle and gridColumnStyle props [#158](https://github.com/hoangnm/react-native-week-view/pull/158)
 ## 0.8.1 - 2021-12-30